# docs/swagger.yaml
definitions:
  ds.AcidSolubleSample:
    properties:
      deleted:
        type: boolean
      description:
        type: string
      formula:
        type: string
      id:
        type: integer
      image_url:
        type: string
      relative_molecular_mass:
        type: number
      stoichiometric_coefficient:
        type: number
      title:
        type: string
    type: object
  ds.ExperimentSample:
    properties:
      acid_soluble_sample:
        $ref: '#/definitions/ds.AcidSolubleSample'
      evolved_gas_volume:
        type: number
      experiment_id:
        type: integer
      impurity_fraction_experiment:
        $ref: '#/definitions/ds.ImpurityFractionExperiment'
      mass_fraction_percentage:
        type: number
      sample_id:
        type: integer
      sample_mass:
        type: number
    type: object
  ds.ImpurityFractionExperiment:
    properties:
      created_at:
        type: string
      creator:
        $ref: '#/definitions/ds.User'
      creator_id:
        type: string
      finished_at:
        type: string
      formed_at:
        type: string
      id:
        type: integer
      moderator:
        $ref: '#/definitions/ds.User'
      moderator_id:
        type: string
      molar_volume:
        type: number
      status:
        type: string
    type: object
  ds.User:
    properties:
      id:
        type: string
      is_moderator:
        type: boolean
      login:
        type: string
      password:
        type: string
    type: object
  dto.AddSampleToExperiment:
    properties:
      created_at:
        type: string
      creator_login:
        type: string
      experiment_id:
        type: integer
      sample_id:
        type: integer
    type: object
  dto.DeleteExperimentSampleResp:
    properties:
      creator_login:
        type: string
      experiment_id:
        type: integer
      sample_id:
        type: integer
    type: object
...
  dto.UpdateExperimentSampleReq:
    properties:
      evolved_gas_volume:
        type: number
      sample_mass:
        type: number
    type: object
...
info:
  contact: {}
paths:
...
  /impurity-experiments/{id}/soluble-samples/{sample_id}:
    delete:
      description: Удаляет связь образца и исследования
      parameters:
      - description: ID исследования
        in: path
        name: id
        required: true
        type: integer
      - description: ID образца
        in: path
        name: sample_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Обновленное исследование
          schema:
            $ref: '#/definitions/dto.DeleteExperimentSampleResp'
        "400":
          description: Неверные ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить образец из исследования
      tags:
      - experiments-samples
    put:
      consumes:
      - application/json
      description: Обновляет параметры образца в конкретном исследовании
      parameters:
      - description: ID исследования
        in: path
        name: id
        required: true
        type: integer
      - description: ID образца
        in: path
        name: sample_id
        required: true
        type: integer
      - description: Новые данные
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateExperimentSampleReq'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленные данные
          schema:
            $ref: '#/definitions/ds.ExperimentSample'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить данные образца в исследовании
      tags:
      - experiments-samples
...
  /soluble-samples:
    get:
      description: Возвращает все образцы или фильтрует по названию
      parameters:
      - description: Название образца для поиска
        in: query
        name: search_sample
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список образцов
          schema:
            items:
              $ref: '#/definitions/ds.AcidSolubleSample'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список образцов
      tags:
      - soluble-samples
    post:
      consumes:
      - application/json
      description: Создает новый образец и возвращает его данные
      parameters:
      - description: Данные нового образца
        in: body
        name: sample
        required: true
        schema:
          $ref: '#/definitions/ds.AcidSolubleSample'
      produces:
      - application/json
      responses:
        "201":
          description: Созданный образец
          schema:
            $ref: '#/definitions/ds.AcidSolubleSample'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать новый образец
      tags:
      - soluble-samples
  /soluble-samples/{id}:
    delete:
      description: Выполняет логическое удаление образца по ID
      parameters:
      - description: ID образца
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Статус удаления
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Образец не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить образец
      tags:
      - soluble-samples
    get:
      description: Возвращает информацию о образце по её идентификатору
      parameters:
      - description: ID образца
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные образца
          schema:
            $ref: '#/definitions/ds.AcidSolubleSample'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Образец не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить образец по ID
      tags:
      - soluble-samples
    put:
      consumes:
      - application/json
      description: Обновляет информацию о образце по ID
      parameters:
      - description: ID образца
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные образца
        in: body
        name: sample
        required: true
        schema:
          $ref: '#/definitions/ds.AcidSolubleSample'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленный образец
          schema:
            $ref: '#/definitions/ds.AcidSolubleSample'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Образец не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить данные образца
      tags:
      - soluble-samples
  /soluble-samples/{id}/experiments/draft:
    post:
      description: Добавляет образец в черновик исследования пользователя
      parameters:
      - description: ID образца
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Исследование с добавленным образцом
          schema:
            $ref: '#/definitions/dto.AddSampleToExperiment'
        "201":
          description: Создано новое исследование
          schema:
            $ref: '#/definitions/dto.AddSampleToExperiment'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Образец не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Образец уже в исследовании
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить образец в черновик исследования
      tags:
      - soluble-samples
  /soluble-samples/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение для образца и возвращает обновленные данные
      parameters:
      - description: ID образца
        in: path
        name: id
        required: true
        type: integer
      - description: Изображение образца
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Статус загрузки и данные образца
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный запрос или файл
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Образец не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Загрузить изображение для образца
      tags:
      - soluble-samples
...
swagger: "2.0"